+++ plugins/exists/test/index.js
@@ -3,30 +3,30 @@ const test = require('tape');
 
 test('Real relative file exists', (assert) => {
   const module = require('../lib');
-  const testFolder = './plugins/exists/test/fixtures/exists.txt';
+  const testPath = './plugins/exists/test/fixtures/exists.txt';
 
-  module.folderExists(testFolder)
+  module.pathExists(testPath)
     .then(() => {
       assert.pass('Resolved promise is returned');
       assert.end();
     })
     .catch(() => {
-      assert.fail(`File system is missing folder (${testFolder})`);
+      assert.fail(`File system is missing folder (${testPath})`);
       assert.end();
     });
 });
 
 test('Real relative folder exists', (assert) => {
   const module = require('../lib');
-  const testFolder = './plugins/exists/test/fixtures';
+  const testPath = './plugins/exists/test/fixtures';
 
-  module.folderExists(testFolder)
+  module.pathExists(testPath)
     .then(() => {
       assert.pass('Resolved promise is returned');
       assert.end();
     })
     .catch(() => {
-      assert.fail(`File system is missing folder (${testFolder})`);
+      assert.fail(`File system is missing folder (${testPath})`);
       assert.end();
     });
 });
@@ -34,15 +34,15 @@ test('Real relative folder exists', (assert) => {
 test('Real absolute file exists', (assert) => {
   const module = require('../lib');
   const path = require('path');
-  const testFolder = path.join(__dirname, './fixtures/exists.txt');
+  const testPath = path.join(__dirname, './fixtures/exists.txt');
 
-  module.folderExists(testFolder)
+  module.pathExists(testPath)
     .then((verifiedPath) => {
-      assert.equal(verifiedPath, testFolder, 'Resolved path matches');
+      assert.equal(verifiedPath, testPath, 'Resolved path matches');
       assert.end();
     })
     .catch(() => {
-      assert.fail(`File system is missing folder (${testFolder})`);
+      assert.fail(`File system is missing folder (${testPath})`);
       assert.end();
     });
 });
@@ -50,15 +50,15 @@ test('Real absolute file exists', (assert) => {
 test('Real absolute folder exists', (assert) => {
   const module = require('../lib');
   const path = require('path');
-  const testFolder = path.join(__dirname, './fixtures');
+  const testPath = path.join(__dirname, './fixtures');
 
-  module.folderExists(testFolder)
+  module.pathExists(testPath)
     .then((verifiedPath) => {
-      assert.equal(verifiedPath, testFolder, 'Resolved path matches');
+      assert.equal(verifiedPath, testPath, 'Resolved path matches');
       assert.end();
     })
     .catch(() => {
-      assert.fail(`File system is missing folder (${testFolder})`);
+      assert.fail(`File system is missing folder (${testPath})`);
       assert.end();
     });
 });
@@ -66,11 +66,11 @@ test('Real absolute folder exists', (assert) => {
 test('Fake absolute path does not exists', (assert) => {
   const module = require('../lib');
   const path = require('path');
-  const testFolder = path.join(__dirname, './fixtures/fakeFolder');
+  const testPath = path.join(__dirname, './fixtures/fakeFolder');
 
-  module.folderExists(testFolder)
+  module.pathExists(testPath)
     .then(() => {
-      assert.fail(`File system found a fake folder (${testFolder})`);
+      assert.fail(`File system found a fake folder (${testPath})`);
       assert.end();
     })
     .catch((error) => {