+++ plugins/rename/lib/index.js
@@ -9,11 +9,17 @@ exports.register = (server, options, next) => {
     config: {
       handler: (request, reply) => {
         const filenames = request.payload.filenames;
-        require('../../exists/lib').folderExists(request.payload.source_folder)
+        require('../../exists/lib').pathExists(request.payload.source_folder)
           .then(() => require('./filenames').getFutureFilenames(request.payload.prefix, filenames.length))
           .then((futureFilenames) => {
-            require('./rename').renamePaths(request.payload.source_folder, filenames, futureFilenames.filenames);
-            reply(futureFilenames.xml);
+            require('./rename').renamePaths(
+              request.payload.source_folder,
+              filenames,
+              futureFilenames.filenames,
+              {
+                renameAssociated: request.payload.rename_associated,
+              });
+            reply({ xml: futureFilenames.xml });
           })
           .catch((error) => {
             let boomError;
@@ -26,13 +32,24 @@ exports.register = (server, options, next) => {
             reply(boomError);
           });
       },
-      tags: ['api'],
+      tags: ['api', 'plugin', 'v1'],
       validate: {
         payload: {
-          filenames: joi.array().items(joi.string().regex(/^[-\w^&'@{}[\],$=!#().%+~ ]+$/)).min(1).max(80).required(),
-          prefix: joi.string().isoDate().required(),
-          source_folder: joi.string().trim().required(),
-          target_folder: joi.string().trim(),
+          filenames: joi.array().items(joi.string().regex(/^[-\w^&'@{}[\],$=!#().%+~ ]+$/))
+            .min(1).max(80).required().example('["DSC01229.JPG"]'),
+          prefix: joi.string().isoDate().required().example('2016-12-31'),
+          source_folder: joi.string().trim().required().example('/public/todo/'),
+          rename_associated: joi.boolean().default(false)
+            .description(`JPG and RAW or video and still image are common ` +
+              `associated pairs that should rename together`),
+        },
+      },
+      response: {
+        schema: {
+          xml: joi.string().required().regex(/<photo\b[^>]*>(.*?)<\/photo>/)
+          .example(`<photo id="1"><filename>2016-04-05-37.jpg</filename></photo>` +
+            `<photo id="2"><filename>2016-04-05-64.jpg</filename></photo>` +
+            `<photo id="3"><filename>2016-04-05-90.jpg</filename></photo>`),
         },
       },
     },